+++ ContentPatcher/release-notes.md
@@ -2,9 +2,9 @@
 
 # Release notes
 ## 1.5.1
-* Fixed issues in Content Patcher 1.5:
-  * error setting `EditData` entries to `null`;
-  * error using tokens in `Enabled` field.
+* Added token support in `EditData` keys (in addition to values).
+* Fixed error setting `EditData` entries to `null` since 1.5.
+* Fixed error using tokens in `Enabled` field since 1.5.
 
 ## 1.5
 * Added support for dynamic tokens defined by the modder.